fix: resolve all 413 pre-existing mypy errors across 84 files
- Add type annotations to untyped functions (128 no-untyped-def)
- Add explicit casts for numpy floating/arg-type mismatches (55 arg-type)
- Fix no-any-return with float()/bool() casts (34 no-any-return)
- Add type: ignore for untyped imports (scipy, psutil) (16 import-untyped)
- Fix dict-item errors in JSON schema examples (16 dict-item)
- Add overloaded spawn() signature on SeededRNG (return-value)
- Rename duplicate RankedOption to CoherenceRankedOption (no-redef)
- Add bidirected edge exclusion filter to E-value computation
- Add gap-closing tests: bidirected exclusion, budget timeout, EVPI
  with goal_constraints, confidence source end-to-end
- Black-format all touched files

+++ src/services/belief_updater.py
@@ -17,7 +17,7 @@ import logging
 from typing import Dict, List
 
 import numpy as np
-from scipy import stats
+from scipy import stats  # type: ignore[import-untyped]
 
 from src.models.phase1_models import (
     CounterfactualQuery,
@@ -134,9 +134,7 @@ class BeliefUpdater:
                 "avg_uncertainty_before": np.mean(
                     list(current_beliefs.uncertainty_estimates.values())
                 ),
-                "avg_uncertainty_after": np.mean(
-                    list(updated_uncertainty.values())
-                ),
+                "avg_uncertainty_after": np.mean(list(updated_uncertainty.values())),
             },
         )
 
@@ -362,7 +360,7 @@ class BeliefUpdater:
         """
         # Compute from uncertainty estimates
         avg_uncertainty = np.mean(list(beliefs.uncertainty_estimates.values()))
-        return avg_uncertainty * 2.0
+        return float(avg_uncertainty * 2.0)
 
     def _identify_top_priorities(
         self,
@@ -414,14 +412,10 @@ class BeliefUpdater:
         top_priorities = self._identify_top_priorities(beliefs, top_n=3)
 
         if len(top_priorities) >= 1:
-            insights.append(
-                f"You strongly prioritize {top_priorities[0]} in your decisions"
-            )
+            insights.append(f"You strongly prioritize {top_priorities[0]} in your decisions")
 
         if len(top_priorities) >= 2:
-            insights.append(
-                f"You value {top_priorities[0]} over {top_priorities[1]}"
-            )
+            insights.append(f"You value {top_priorities[0]} over {top_priorities[1]}")
 
         # Check uncertainty levels
         avg_uncertainty = np.mean(list(beliefs.uncertainty_estimates.values()))
